FeatureCollection add type parameter T<:Feature and features field
The type parameter T is taken from peeking at the first feature if it exists. This is needed to make iterating over the features ini a featurecollection type stable.

The features field was added after I noticed that the change above hardly helped iteration performance. On every `iterate` call the features had to be taken out of the JSON3.Object, which cannot be inferred. So it is better to keep them in the object, since they are a view on the same tape anyway.